# docs/swagger.yaml
basePath: /
definitions:
...
  models.Auth:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.AuthorizationDTO:
    properties:
      token:
        type: string
    type: object
  models.CategoriesGroup:
    properties:
      category_id:
        type: integer
      category_name:
        type: string
      items:
        items:
          $ref: '#/definitions/models.ItemDTO'
        type: array
    type: object
  models.CategoryDTO:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.Created:
    properties:
      inserted_id:
        type: integer
    type: object
  models.ErrorMapDTO:
    properties:
      issue: {}
    type: object
  models.ItemDTO:
    properties:
      category_id:
        type: integer
      id:
        type: integer
      image_url:
        type: string
      name:
        type: string
      note:
        type: string
    type: object
  models.ItemFormDTO:
    properties:
      category_id:
        type: integer
      file: {}
      id:
        type: integer
      name:
        type: string
      note:
        type: string
    type: object
...
  models.UserDTO:
    properties:
      email:
        type: string
      lastname:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
info:
  contact: {}
  description: shopping list site APIs
  title: shoppingify-server APIs
  version: "1.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Login to a new user permanent
      parameters:
      - description: auth credentials
        in: body
        name: auth
        required: true
        schema:
          $ref: '#/definitions/models.Auth'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuthorizationDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorMapDTO'
        "500":
          description: ""
      summary: Login as saved user
      tags:
      - accounts
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Registers a new user permanent
      parameters:
      - description: auth info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuthorizationDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorMapDTO'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.ErrorMapDTO'
        "500":
          description: ""
      summary: Registers a user
      tags:
      - accounts
  /api/auth/visitor-register:
    post:
      consumes:
      - application/json
      description: Login to a new user permanent
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AuthorizationDTO'
        "500":
          description: ""
      summary: Login as saved user
      tags:
      - accounts
  /api/categories:
    get:
      consumes:
      - application/json
      description: Get categories. Pagination is able
      parameters:
      - description: items to take in query
        in: query
        name: take
        type: integer
      - description: items to skip in query
        in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CategoryDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorMapDTO'
        "500":
          description: ""
      summary: Get categories
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Creates a new category. It needs to have an unique name
      parameters:
      - description: categorty info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.CategoryDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Created'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorMapDTO'
        "500":
          description: ""
      summary: Creates a new category
      tags:
      - categories
  /api/categories/by-name:
    get:
      consumes:
      - application/json
      description: Get categories by name. Pagination is able
      parameters:
      - description: category search query
        in: query
        name: q
        required: true
        type: string
      - description: items to take in query
        in: query
        name: take
        type: integer
      - description: items to skip in query
        in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CategoryDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorMapDTO'
        "500":
          description: ""
      summary: Get categories by name
      tags:
      - categories
  /api/items:
    get:
      consumes:
      - application/json
      description: Get items by category groups. Pagination is available
      parameters:
      - description: items to take in query
        in: query
        name: take
        type: integer
      - description: items to skip in query
        in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CategoriesGroup'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorMapDTO'
        "500":
          description: ""
      summary: Get items by category groups
      tags:
      - items
    post:
      consumes:
      - multipart/form-data
      description: Creates a new item. It needs to have an unique name
      parameters:
      - description: item info
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/models.ItemFormDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Created'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorMapDTO'
        "500":
          description: ""
      summary: Creates a new item
      tags:
      - items
  /api/items/{id}:
    get:
      consumes:
      - application/json
      description: Get an item by id
      parameters:
      - description: item ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ItemDTO'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorMapDTO'
        "500":
          description: ""
      summary: Get an item by id
      tags:
      - items
...
swagger: "2.0"